test: System test for notifications
