+++ app/views/notifications/_notification.html.erb
@@ -1,4 +1,4 @@
-<div role="list" class="w-full max-w-xl mx-auto rounded-lg overflow-hidden border border-gray-200 divide-y divide-gray-200 mt-3">
+<div role="list" id="<%= dom_id(notification) %>" class="w-full max-w-xl mx-auto rounded-lg overflow-hidden border border-gray-200 divide-y divide-gray-200 mt-3">
   <%= link_to notification_path(notification), class: "w-full" do %>
     <div class="relative bg-white py-5 px-6 hover:bg-gray-50 focus-within:ring-2 focus-within:ring-inset focus-within:ring-blue-600">
       <div class="flex justify-between space-x-3">
