+++ test/system/login_test.rb
@@ -7,14 +7,4 @@ class LoginTest < ApplicationSystemTestCase
 
     assert page.has_button?("Sign out")
   end
-
-  private
-
-  def login_as(user)
-    visit new_user_session_path
-    fill_in "E-mail", with: user.email
-    fill_in "Password", with: "password"
-
-    click_button "Sign in"
-  end
 end
